+++ Assets/Data/item_status.json.meta
@@ -1,7 +1,6 @@
 fileFormatVersion: 2
-guid: 04693201b946aac419fa322ee5417ec6
-folderAsset: yes
-DefaultImporter:
+guid: 3cb469e881224654c9b833efe7bc6949
+TextScriptImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
