+++ Assets/Scripts/Audio/AudioName.cs
@@ -1,11 +1,12 @@
-﻿using UnityEngine;
-
-/// <summary>
+﻿/// <summary>
 /// 内部逻辑音频索引名到外部物理音频文件名的映射
 /// </summary>
-public class AudioNames {
+public class AudioName {
 
-  public const string BGM = "";
+  // BGM
+  public const string BGM = "in_case_that___";
+
+  // SFX
   public const string SettingClick = "";
   public const string PlayerMove = "a_bat_icespread";
   public const string PlayerJump = "";
