+++ Assets/Scripts/Backpack/Backpack.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 48d70abde2278b840a088a1c40ba4ad0
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
