+++ Assets/Scripts/Backpack/Item.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 53f5a4e96167c9c4ab7a9bbce7dd4924
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
