+++ Assets/Scripts/Backpack/ItemDef.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 0a9523f975b63634ca3c814ae337545e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
