+++ Assets/Scripts/Backpack/ItemDefConfigFile.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 0e5331abe34ff8b4c88b217d8900204b
+guid: 44664c7dcbcc26143bf55dc3ed387b89
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
