+++ Assets/Scripts/Backpack/ItemName.cs
@@ -1,0 +1,11 @@
+﻿/// <summary>
+/// 道具名单 (可能没用)
+/// </summary>
+public class ItemName
+{
+  public const string KusaSeed = "whatsoever seed";
+  public const string HanaSeed = "not a seed";
+  public const string LumineDot = "rwkk";
+  public const string LumineRock = "can can need";
+  public const string PonponBakudan = "kureeeee!";
+}
