+++ Assets/Scripts/Backpack/ItemName.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 32798b261d7aec14e9f501b5a4c97912
+guid: ed85b804d0985f349a580115f935091d
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
