+++ Assets/Scripts/Backpack/ItemType.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: faa8721e14d805f44ac1d2732bcc63c3
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
