+++ Assets/Scripts/EnemyAi/EnemyParameter.cs
@@ -1,0 +1,4 @@
+﻿public class EnemyParameter
+{
+
+}
