+++ Assets/Scripts/EnemyAi/EnemyParameter.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b0e1f98b49cb33540936e683e1ffa7e9
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
