+++ Assets/Scripts/EnemyAi/IState.cs
@@ -1,8 +1,4 @@
-﻿using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-/// <summary>
+﻿/// <summary>
 /// 所有状态的接口
 /// </summary>
 public interface IState {
